requirements.txt
# Multi-Agent Healthcare Chatbot Requirements

# Core dependencies
autogen==0.7
openai==1.64.0
python-dotenv==1.1.0

# Web demo dependencies
streamlit>=1.28.0

# Optional dependencies for enhanced functionality
nbconvert>=7.0.0
jupyter>=1.0.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
